style: 💄 Close button, Alert ribbon
Updated styling to match design. Also added story for Close button...
The new style for close button needed to add a child to the button to
handle icon style. This was so the button element could be used for a
larger touch area on mobile.

BREAKING CHANGE: 🧨 Close button need to have a i-elemented added as a child to the button
tag in order to display the cross icon.