refactor: 💡 AlertRibbon
Refactoring the Alert component so it's named AlertRibbon. Keeping Alert
as an exported constant to be backwards compatible. Moving to style this
component based on the class name "alert ribbon" rather than using the
role attribute due to that we need to have the flexiblity to change the
role to "status" or "dialog". Also hiding the icon with aria-hidden
since it's just decorative and the alert message is the main focus.
Changed header not to fall back to h3 but rather span with header class
for styling. Added onClose event. Fixed test after component change
since they broke. Also removing the i element from icons since they
should only contain the svg icon and not the i element.

BREAKING CHANGE: 🧨 Renaming Alert export to AlertRibbon, changing styling form
[role="alert"] to .alert-ribbon

✅ Closes: 659