fix: 🐛 move transition to hover
The transition being set directly in the mixin caused some awkward
visual effects when using e.g. visibility: hidden on a parent element.
Moving the transition styling to appear only on hover effect, which is
the intended use of it, keeps the visual transition effect when needed,
but does not give awkward visual effects when using visibility:hidden.