fix: 🐛 fix context menu position in relative parent
When context menu was put in a parent HTML element, it was displayed in
a wrong position due to top and left coordinates being calculated from
the whole window, not from parent coordinate system. This change
introduces an overlay in order to mitigate the issue in the simplest
possible manner. The solution was inspired by material angular library
tooltip component implementation.

BREAKING CHANGE: 🧨 -

✅ Closes: -